Move entityCompatibleQuery to metabase/api/utils/entity-compatible-query.ts
