+++ frontend/src/metabase/admin/permissions/permissions.ts
@@ -20,8 +20,8 @@ import {
 } from "metabase/admin/permissions/utils/graph";
 import { getGroupFocusPermissionsUrl } from "metabase/admin/permissions/utils/urls";
 import { databaseApi, permissionApi } from "metabase/api";
+import { entityCompatibleQuery } from "metabase/api/utils/entity-compatible-query";
 import { type ErrorPayload, getErrorMessage } from "metabase/api/utils/errors";
-import { entityCompatibleQuery } from "metabase/entities/utils";
 import {
   PLUGIN_ADVANCED_PERMISSIONS,
   PLUGIN_DATA_PERMISSIONS,
