+++ frontend/src/metabase/api/utils/index.ts
@@ -1,3 +1,4 @@
+export * from "./entity-compatible-query";
 export * from "./errors";
 export * from "./hydrate-legacy-entities";
 export * from "./settings";
