+++ frontend/src/metabase/dashboard/actions/cards-typed.ts
@@ -3,7 +3,7 @@ import { t } from "ttag";
 import _ from "underscore";
 
 import { cardApi } from "metabase/api";
-import { entityCompatibleQuery } from "metabase/entities/utils";
+import { entityCompatibleQuery } from "metabase/api/utils/entity-compatible-query";
 import { loadMetadataForCard } from "metabase/questions/actions";
 import { createThunkAction } from "metabase/redux";
 import type { Dispatch, GetState } from "metabase/redux/store";
