+++ frontend/src/metabase/dashboard/actions/revisions.ts
@@ -1,9 +1,9 @@
 import { revisionApi } from "metabase/api";
+import { entityCompatibleQuery } from "metabase/api/utils/entity-compatible-query";
 import {
   fetchDashboard,
   fetchDashboardCardData,
 } from "metabase/dashboard/actions";
-import { entityCompatibleQuery } from "metabase/entities/utils";
 import { createThunkAction } from "metabase/redux";
 import { REVERT_TO_REVISION } from "metabase/redux/dashboard";
 import type { DashboardId, Revision } from "metabase-types/api";
