+++ frontend/src/metabase/dashboard/actions/save.ts
@@ -2,7 +2,7 @@ import { assocIn, dissocIn, getIn } from "icepick";
 import _ from "underscore";
 
 import { cardApi, dashboardApi } from "metabase/api";
-import { entityCompatibleQuery } from "metabase/entities/utils";
+import { entityCompatibleQuery } from "metabase/api/utils/entity-compatible-query";
 import { clickBehaviorIsValid } from "metabase/parameters/utils/click-behavior";
 import { createThunkAction } from "metabase/redux";
 import { UPDATE_DASHBOARD_AND_CARDS } from "metabase/redux/dashboard";
