+++ frontend/src/metabase/redux/metadata.js
@@ -10,7 +10,7 @@ import {
   segmentApi,
   tableApi,
 } from "metabase/api";
-import { entityCompatibleQuery } from "metabase/entities/utils";
+import { entityCompatibleQuery } from "metabase/api/utils/entity-compatible-query";
 import { isProduction } from "metabase/env";
 import { createThunkAction } from "metabase/redux";
 import { fetchTableMetadataAndForeignKeys } from "metabase/redux/tables";
